KVM: selftests: Fix gmem conversion tests for multiple vCPUs
Currently the private_mem_conversions_test crashes if invoked with the
-n <num_vcpus> option without also specifying multiple memslots via -m.

This is because the current implementation assumes -m is specified and
always sets up the per-vCPU memory with a dedicated memslot for each
vCPU. When -m is not specified, the test skips setting up
memslots/memory for secondary vCPUs.

The current code does seem to try to handle using a single memslot for
multiple vCPUs in some places, e.g. the call-site, but
test_mem_conversions() is missing the important bit of sizing the single
memslot appropriately to handle all the per-vCPU memory. Implement that
handling.

Signed-off-by: Michael Roth <michael.roth@amd.com>